Add ResultSet mixin
