+++ pyathena/common.py
@@ -209,3 +209,78 @@ class BaseCursor(with_metaclass(ABCMeta, object)):
 
     def __exit__(self, exc_type, exc_val, exc_tb):
         self.close()
+
+
+class WithResultSet(object):
+
+    def __init__(self):
+        self._query_id = None
+        self._result_set = None
+
+    def _reset_state(self):
+        self._query_id = None
+        if self._result_set and not self._result_set.is_closed:
+            self._result_set.close()
+        self._result_set = None
+
+    @property
+    def has_result_set(self):
+        return self._result_set is not None
+
+    @property
+    def description(self):
+        if not self.has_result_set:
+            return None
+        return self._result_set.description
+
+    @property
+    def query_id(self):
+        return self._query_id
+
+    @property
+    def query(self):
+        if not self.has_result_set:
+            return None
+        return self._result_set.query
+
+    @property
+    def state(self):
+        if not self.has_result_set:
+            return None
+        return self._result_set.state
+
+    @property
+    def state_change_reason(self):
+        if not self.has_result_set:
+            return None
+        return self._result_set.state_change_reason
+
+    @property
+    def completion_date_time(self):
+        if not self.has_result_set:
+            return None
+        return self._result_set.completion_date_time
+
+    @property
+    def submission_date_time(self):
+        if not self.has_result_set:
+            return None
+        return self._result_set.submission_date_time
+
+    @property
+    def data_scanned_in_bytes(self):
+        if not self.has_result_set:
+            return None
+        return self._result_set.data_scanned_in_bytes
+
+    @property
+    def execution_time_in_millis(self):
+        if not self.has_result_set:
+            return None
+        return self._result_set.execution_time_in_millis
+
+    @property
+    def output_location(self):
+        if not self.has_result_set:
+            return None
+        return self._result_set.output_location
