+++ pyathena/pandas_cursor.py
@@ -12,6 +12,7 @@ from future.utils import raise_from
 
 from pyathena import DataError
 from pyathena.common import CursorIterator
+from pyathena.common import WithResultSet
 from pyathena.cursor import BaseCursor
 from pyathena.error import NotSupportedError, OperationalError, ProgrammingError
 from pyathena.model import AthenaQueryExecution
@@ -20,7 +21,7 @@ from pyathena.util import retry_api_call
 _logger = logging.getLogger(__name__)
 
 
-class PandasCursor(BaseCursor):
+class PandasCursor(BaseCursor, WithResultSet):
 
     _pattern_output_location = re.compile(r'^s3://(?P<bucket>[a-zA-Z0-9.\-_]+)/(?P<key>.+)$')
 
@@ -37,65 +38,6 @@ class PandasCursor(BaseCursor):
             's3', region_name=self.connection.region_name, **self.connection._kwargs)
         self._arraysize = arraysize
 
-        self._query_id = None
-        self._query_execution = None
-
-    @property
-    def has_query_execution(self):
-        return self._query_execution is not None
-
-    @property
-    def query_id(self):
-        return self._query_id
-
-    @property
-    def query(self):
-        if not self.has_query_execution:
-            return None
-        return self._query_execution.query
-
-    @property
-    def state(self):
-        if not self.has_query_execution:
-            return None
-        return self._query_execution.state
-
-    @property
-    def state_change_reason(self):
-        if not self.has_query_execution:
-            return None
-        return self._query_execution.state_change_reason
-
-    @property
-    def completion_date_time(self):
-        if not self.has_query_execution:
-            return None
-        return self._query_execution.completion_date_time
-
-    @property
-    def submission_date_time(self):
-        if not self.has_query_execution:
-            return None
-        return self._query_execution.submission_date_time
-
-    @property
-    def data_scanned_in_bytes(self):
-        if not self.has_query_execution:
-            return None
-        return self._query_execution.data_scanned_in_bytes
-
-    @property
-    def execution_time_in_millis(self):
-        if not self.has_query_execution:
-            return None
-        return self._query_execution.execution_time_in_millis
-
-    @property
-    def output_location(self):
-        if not self.has_query_execution:
-            return None
-        return self._query_execution.output_location
-
     def _parse_output_location(self, output_location):
         match = self._pattern_output_location.search(output_location)
         if match:
@@ -106,16 +48,17 @@ class PandasCursor(BaseCursor):
     def close(self):
         pass
 
-    def _reset_state(self):
-        self._query_id = None
-        self._query_execution = None
-
     def execute(self, operation, parameters=None):
         self._reset_state()
         self._query_id = self._execute(operation, parameters)
-        self._query_execution = self._poll(self._query_id)
-        if self._query_execution.state != AthenaQueryExecution.STATE_SUCCEEDED:
-            raise OperationalError(self._query_execution.state_change_reason)
+        query_execution = self._poll(self._query_id)
+        if query_execution.state == AthenaQueryExecution.STATE_SUCCEEDED:
+            self._result_set = AthenaResultSet(
+                self._connection, self._converter, query_execution, self._arraysize,
+                self.retry_exceptions, self.retry_attempt, self.retry_multiplier,
+                self.retry_max_delay, self.retry_exponential_base)
+        else:
+            raise OperationalError(query_execution.state_change_reason)
         return self
 
     def executemany(self, operation, seq_of_parameters):
@@ -127,8 +70,8 @@ class PandasCursor(BaseCursor):
         self._cancel(self._query_id)
 
     def as_pandas(self):
-        if not self.has_query_execution:
-            raise ProgrammingError('No query execution.')
+        if not self.has_result_set:
+            raise ProgrammingError('No result set.')
         bucket, key = self._parse_output_location(self.output_location)
         try:
             response = retry_api_call(self._client.get_object,
